+++ Tests/GRDBTests/DatabaseRegionTests.swift
@@ -265,19 +265,19 @@ class DatabaseRegionTests : GRDBTestCase {
                 let statement = try db.makeSelectStatement("SELECT foo.name FROM FOO JOIN BAR ON fooId = foo.id")
                 let expectedRegion = DatabaseRegion(table: "foo", columns: ["name", "id"])
                     .union(DatabaseRegion(table: "bar", columns: ["fooId"]))
-                XCTAssertEqual(statement.fetchedRegion, expectedRegion)
-                XCTAssertEqual(statement.fetchedRegion.description, "bar(fooId),foo(id,name)")
+                XCTAssertEqual(statement.databaseRegion, expectedRegion)
+                XCTAssertEqual(statement.databaseRegion.description, "bar(fooId),foo(id,name)")
             }
             do {
                 let statement = try db.makeSelectStatement("SELECT COUNT(*) FROM foo")
                 if sqlite3_libversion_number() < 3019000 {
                     let expectedRegion = DatabaseRegion.fullDatabase
-                    XCTAssertEqual(statement.fetchedRegion, expectedRegion)
-                    XCTAssertEqual(statement.fetchedRegion.description, "full database")
+                    XCTAssertEqual(statement.databaseRegion, expectedRegion)
+                    XCTAssertEqual(statement.databaseRegion.description, "full database")
                 } else {
                     let expectedRegion = DatabaseRegion(table: "foo")
-                    XCTAssertEqual(statement.fetchedRegion, expectedRegion)
-                    XCTAssertEqual(statement.fetchedRegion.description, "foo(*)")
+                    XCTAssertEqual(statement.databaseRegion, expectedRegion)
+                    XCTAssertEqual(statement.databaseRegion.description, "foo(*)")
                 }
             }
         }
@@ -295,117 +295,117 @@ class DatabaseRegionTests : GRDBTestCase {
             
             do {
                 let request = Record.all()
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)")
             }
             do {
                 let request = Record.filter(Column("a") == 1)
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)")
             }
             do {
                 let request = Record.filter(Column("id") >= 1)
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)")
             }
             
             do {
                 let request = Record.filter((Column("id") == 1) || (Column("a") == "foo"))
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)")
             }
 
             // No rowId
             
             do {
                 let request = Record.filter(Column("id") == nil)
-                try XCTAssertEqual(request.fetchedRegion(db).description, "empty")
+                try XCTAssertEqual(request.databaseRegion(db).description, "empty")
             }
 
             do {
                 let request = Record.filter(Column("id") === nil)
-                try XCTAssertEqual(request.fetchedRegion(db).description, "empty")
+                try XCTAssertEqual(request.databaseRegion(db).description, "empty")
             }
             
             do {
                 let request = Record.filter(nil == Column("id"))
-                try XCTAssertEqual(request.fetchedRegion(db).description, "empty")
+                try XCTAssertEqual(request.databaseRegion(db).description, "empty")
             }
             
             do {
                 let request = Record.filter(nil === Column("id"))
-                try XCTAssertEqual(request.fetchedRegion(db).description, "empty")
+                try XCTAssertEqual(request.databaseRegion(db).description, "empty")
             }
             
             do {
                 let request = Record.filter((Column("id") == 1) && (Column("id") == 2))
-                try XCTAssertEqual(request.fetchedRegion(db).description, "empty")
+                try XCTAssertEqual(request.databaseRegion(db).description, "empty")
             }
             do {
                 let request = Record.filter(key: 1).filter(key: 2)
-                try XCTAssertEqual(request.fetchedRegion(db).description, "empty")
+                try XCTAssertEqual(request.databaseRegion(db).description, "empty")
             }
 
             // Single rowId
             
             do {
                 let request = Record.filter(Column("id") == 1)
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1]")
             }
             do {
                 let request = Record.filter(Column("id") === 1)
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1]")
             }
             do {
                 let request = Record.filter(Column("id") == 1 && Column("a") == "foo")
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1]")
             }
             do {
                 let request = Record.filter(Column.rowID == 1)
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1]")
             }
             do {
                 let request = Record.filter(1 == Column("id"))
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1]")
             }
             do {
                 let request = Record.filter(1 === Column("id"))
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1]")
             }
             do {
                 let request = Record.filter(1 === Column.rowID)
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1]")
             }
             do {
                 let request = Record.filter(key: 1)
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1]")
             }
             do {
                 let request = Record.filter(key: 1).filter(key: 1)
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1]")
             }
             do {
                 let request = Record.filter(key: 1).filter(Column("a") == "foo")
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1]")
             }
 
             // Multiple rowIds
             
             do {
                 let request = Record.filter(Column("id") == 1 || Column.rowID == 2)
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1,2]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1,2]")
             }
             do {
                 let request = Record.filter((Column("id") == 1 && Column("a") == "foo") || Column.rowID == 2)
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1,2]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1,2]")
             }
             do {
                 let request = Record.filter([1, 2, 3].contains(Column("id")))
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1,2,3]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1,2,3]")
             }
             do {
                 let request = Record.filter([1, 2, 3].contains(Column.rowID))
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1,2,3]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1,2,3]")
             }
             do {
                 let request = Record.filter(keys: [1, 2, 3])
-                try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1,2,3]")
+                try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1,2,3]")
             }
         }
     }
@@ -433,14 +433,14 @@ class DatabaseRegionTests : GRDBTestCase {
                     .including(optional: A.b.filter(key: 2))
                     .including(optional: A.c.filter(keys: [1, 2, 3]))
                 // This test will fail when we are able to improve regions of joined requestt
-                try XCTAssertEqual(request.fetchedRegion(db).description, "a(id,name)[1],b(aid,id,name),c(aid,id,name)")
+                try XCTAssertEqual(request.databaseRegion(db).description, "a(id,name)[1],b(aid,id,name),c(aid,id,name)")
             }
             do {
                 let request = B.filter(key: 1)
                     .including(optional: B.a.filter(key: 2)
                         .including(optional: A.c.filter(keys: [1, 2, 3])))
                 // This test will fail when we are able to improve regions of joined requestt
-                try XCTAssertEqual(request.fetchedRegion(db).description, "a(id,name),b(aid,id,name)[1],c(aid,id,name)")
+                try XCTAssertEqual(request.databaseRegion(db).description, "a(id,name),b(aid,id,name)[1],c(aid,id,name)")
             }
         }
    }
@@ -454,20 +454,20 @@ class DatabaseRegionTests : GRDBTestCase {
             }
             
             let request = Record.filter(keys: [1, 2, 3])
-            try XCTAssertEqual(request.fetchedRegion(db).description, "foo(a,id)[1,2,3]")
+            try XCTAssertEqual(request.databaseRegion(db).description, "foo(a,id)[1,2,3]")
 
             do {
                 let derivedRequest: AnyFetchRequest<Row> = AnyFetchRequest(request)
-                try XCTAssertEqual(derivedRequest.fetchedRegion(db).description, "foo(a,id)[1,2,3]")
+                try XCTAssertEqual(derivedRequest.databaseRegion(db).description, "foo(a,id)[1,2,3]")
             }
             do {
                 let derivedRequest: AdaptedFetchRequest = request.adapted { db in SuffixRowAdapter(fromIndex: 1) }
-                try XCTAssertEqual(derivedRequest.fetchedRegion(db).description, "foo(a,id)[1,2,3]")
+                try XCTAssertEqual(derivedRequest.databaseRegion(db).description, "foo(a,id)[1,2,3]")
             }
             do {
                 // SQL request loses region info
                 let derivedRequest = try SQLRequest(db, request: request)
-                try XCTAssertEqual(derivedRequest.fetchedRegion(db).description, "foo(a,id)")
+                try XCTAssertEqual(derivedRequest.databaseRegion(db).description, "foo(a,id)")
             }
         }
     }
@@ -506,20 +506,20 @@ class DatabaseRegionTests : GRDBTestCase {
             do {
                 let statement = try db.makeSelectStatement("SELECT rowid FROM FOO")
                 let expectedRegion = DatabaseRegion(table: "foo", columns: ["ROWID"])
-                XCTAssertEqual(statement.fetchedRegion, expectedRegion)
-                XCTAssertEqual(statement.fetchedRegion.description, "foo(ROWID)")
+                XCTAssertEqual(statement.databaseRegion, expectedRegion)
+                XCTAssertEqual(statement.databaseRegion.description, "foo(ROWID)")
             }
             do {
                 let statement = try db.makeSelectStatement("SELECT _ROWID_ FROM FOO")
                 let expectedRegion = DatabaseRegion(table: "foo", columns: ["ROWID"])
-                XCTAssertEqual(statement.fetchedRegion, expectedRegion)
-                XCTAssertEqual(statement.fetchedRegion.description, "foo(ROWID)")
+                XCTAssertEqual(statement.databaseRegion, expectedRegion)
+                XCTAssertEqual(statement.databaseRegion.description, "foo(ROWID)")
             }
             do {
                 let statement = try db.makeSelectStatement("SELECT oID FROM FOO")
                 let expectedRegion = DatabaseRegion(table: "foo", columns: ["ROWID"])
-                XCTAssertEqual(statement.fetchedRegion, expectedRegion)
-                XCTAssertEqual(statement.fetchedRegion.description, "foo(ROWID)")
+                XCTAssertEqual(statement.databaseRegion, expectedRegion)
+                XCTAssertEqual(statement.databaseRegion.description, "foo(ROWID)")
             }
         }
     }